Refactor: Reorganized account classes into one module. Modified tests into one account test file.
