+++ tests/test_simulation.py
@@ -1,17 +1,3 @@
-from backtester.brokerage.backtest_account import BacktestCashAccount
-import csv
-
-def test_simulation():
-    # feed in the data
-    acct = BacktestCashAccount()
-    acct.set_cash(1000)
-
-    with open("SPY.csv", newline='') as csv_file:
-        read = csv.reader(csv_file)
-
-    
 
 
 
-if __name__ == "__main__":
-    test_simulation()
